Fixed login issue with custom axios api
The problem that was happpening was that our JWT tokens were only configured to be stored for a mobile app. Our request was failing because it wasn't configured
for tokens to be saved on a web application. I had to update tokenStorage.ts to handle saving tokens for web application. Essentially tokenStorage.ts securley stores the JWT tokens
on the front end retrieves those tokens from storage, attaches them to backend requests, and refreshes them when they expire. Important note for backend work. If thee are any controllers
that need to have JWT authoriization add the line --> before_action :authorize_request. This establishes JWT authorization for that controller.